+++ NEWS.md
@@ -1,4 +1,4 @@
-# glue 1.1.1.9000
+# glue 1.2.0
 
 * The implementation has been tweaked to be slightly faster in most cases.